requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
email-validator==2.2.0
python-multipart
